fix: clarify comments on persistentDataStoreWrapper Init() and remove always-true conditional (#361)
The persistentDataStoreWrapper's Init method is concerned with initializing its underlying store and then setting up
the caching layer. 

The conditional check at the end of the function that guards setting the initialization status
is not necessary, because it's not possible to arrive at the check without either (no error) or (cache is infinite ttl) being true.