chore: Create FDv2 compatible polling data source
Expands the `datasourcev2` package with a polling datasource
implementation that is compatible with our current SDK, but pulls data
updates from a FDv2-compatible source.

It does not support any of the other FDv2-specific features like picking
up from a known state. That will be introduced in later work as we
re-shape the datasource integration on a larger scale.