refactor: streaming/polling processors take DataDestination/StatusReporter interfaces (#191)
This change splits up the monolithic `DataSourceUpdateSink` into two:
1. `DataDestination`: where data goes
2. `StatusReporter`: where statuses go

This makes it easier to mock them for testing, and also represents a
cleaner separation of concerns. These new interfaces are slightly
different than the `DataSourceUpdateSink` for FDV1 vs FDV2.

In the new system, we want to be able to do things like passing a
payload version in `Init`, whereas the old FDV1 code has no such concept.