refactor: extract topographical sort logic into dedicated package (#196)
The existing data source update sink has logic for computing a
topological sort on the SDK's data set. This is ultimately used to
insert data into a persistent store via `Init`.

If we didn't do this, then (since stores are not atomic) evaluations for
a given flag might require a prerequisite or segment that doesn't yet
exist in the store.

This logic is also needed in the FDv2 data system. Since it was an
implementation detail of the existing v1 data sources, I've extracted it into a
dedicated package. 

Finally, this commit modifies the types and methods to use graphing terminology, hoping to make its function more intuitive.