feat: add support for client-side prerequisite events (#201)
This commit updates `AllFlagsState` to track prerequisite evaluations. 

This didn't require modifying the eval module, as it already exposes a
`PrerequisiteEventRecorder` interface with the necessary info.

The `AllFlagsState` public API allows for fetching a flag's details
(`FlagState`) from the top-level `AllFlags` object. This struct now has
prerequisite information exposed.

Additionally when the `AllFlags` is marshaled to JSON, it will now
contain the prerequisite relationships for each flag.