fix: pass parsed versionKey to Tracker (#211)
The AI client wasn't passing the parsed `_ldMeta.versionKey` into newly
created `Tracker`s. The tests didn't catch it because the default value
of strings are `""`, and that's what the tests asserted against.

I've updated the tests to pass in a non-default string to make sure it
is being plumbed through.