feat: pass tracker's config into TrackRequest (#210)
This PR improves the ergonomics of `TrackRequest`. The user callback now
receives a `Config` directly, which can be accessed to provide
parameters for the request.

An alternative would be directly exposing the `Config` as a public
variable of `Tracker`, or forcing the user to instead reference the
config returned from a call to `cfg, tracker := client.Config(..)`.

Those alternatives are not great, because they make the user carry
around the config and manually inject it into the callback. This way,
the callback can be defined wherever and not need to worry about how to
receive the config.