chore: Fix implicit intent handling
Synchronizers operate by building up changes using a change set builder.
This builder is primed with a server-intent, and then updated as
subsequent changes are received.

If the initial response to the synchronizer is `xfer-none`, we still
need to store this intent. However, the intent code should be updated to
show that we are now waiting for pending changes. Otherwise, when those
subsequent changes arrive, the code will suggest to upstream behavior
that we have no changes to process.