chore: Remove data storage out of initializers and synchronizers (#271)
The initializers and synchronizers are now required to return the data
they have retrieved instead of persisting it directly to the data
destination. This provides for a better separation of concerns where the
synchronizers only have to worry about their data formats, and not
persistent.

We do update each initializers or synchronizers with access to the
current selector using the `DataSelector` interface.