feat: Support latest feature flagging semantic conventions. (#292)
Updates the tracing hook to use the latest semantic conventions.

This drops using the semver components provided as part of open
telemetry because that is connected to the otel SDK version being used.
Updating it required updating the other otel components.

I would prefer to decouple what is effectively a few strings from the
versions of the OTEL dependency being used.