fix(experimental): NewScopedClient can now only take one context argument (#309)
Currently, NewScopedClient takes a variadic (zero or more) list of
contexts to initially seed the scoped client. However, this means it's
possible to pass in _zero_ contexts, resulting in an invalid state: a
scoped client with zero contexts.

To avoid this scenario, I propose changing `NewScopedClient` to only
take in one context, which could be a multi-context. This is the
simplest implementation choice and it doesn't block any of the common
use cases I anticipate: adding 1 or N contexts will still be possible
with this syntax.