chore: Use gh-actions/actions/persistent-stores (#319)
The latest consul image incorrectly handles a '$' in the key name. This
is causing failing tests since we use an '$inited' key. Until the consul
change is remedied, we need to pin to a working version of consul.

This is already done in our shared GH action for persistent stores, so I
am updating this SDK to finally use that.