fix: Add `ChangeSet.Collections` to help reduce relay memory usage (#332)
The SDK changes to convert changesets into `[]ldstoretypes.Collection`
for a lot of it's internal workings. Coincidentally, the relay proxy
also needs to do this same conversion when supporting FDv1 formats.

To help with this, we add a `.Collections` method which memoizes the
result for later use, reducing our memory allocation.