fix: Optimize creation of changesets from collections
Previously, we worked to cache the changset -> collections process as it
is used in many different places. There are multiple places where a
changeset is created from a collection. With this change, we take
advantage of this knowledge, and pre-cache the collection that created
the changeset, preventing us from doing the conversion at all.